requirements.txt
ctgan==0.5.1
lightgbm==3.3.2
matplotlib==3.5.2
numpy==1.21.6
optuna==2.10.0
pandas==1.3.5
scikit-learn==1.0.2
scipy==1.7.3
torchviz==0.0.2
tqdm==4.64.0
xgboost==1.3.3
